Refactor recipe generator to enable scraping recipes
It's overly tedious to manually add new recipes, even with the help of
the recipe template generator. This is the start of the work needed to
automatically generate a recipe template based off of a supplied URL.

Scraping the recipe and presenting it in a YAML template will seriously
cut down on the amount of manual work it takes to add recipes I like.
Since most of the good ones come from a handful of sites, I should only
need to implement a couple of scrapers.